# swagger/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "# API Usage\n\nMost actions that can be performed via API can be performed\
    \ via UI. We recommend managing your networks using the official netmaker-ui project.\
    \ However, Netmaker can also be run without the UI, and all functions can be achieved\
    \ via API calls. If your use case requires using Netmaker without the UI or you\
    \ need to do some troubleshooting/advanced configuration, using the API directly\
    \ may help.\n\n# Authentication\n\nAPI calls must be authenticated via a header\
    \ of the format -H “Authorization: Bearer <YOUR_SECRET_KEY>” There are two methods\
    \ to obtain YOUR_SECRET_KEY: 1. Using the masterkey. By default, this value is\
    \ “secret key,” but you should change this on your instance and keep it secure.\
    \ This value can be set via env var at startup or in a config file (config/environments/<\
    \ env >.yaml). See the [Netmaker](https://docs.netmaker.org/index.html) documentation\
    \ for more details. 2. Using a JWT received for a node. This can be retrieved\
    \ by calling the /api/nodes/<network>/authenticate endpoint, as documented below."
  version: "0.24.0"
  title: "Netmaker"
host: "api.demo.netmaker.io"
basePath: "/"
schemes:
- "https"
consumes:
- "application/json"
produces:
- "application/json"
paths:
...
  /api/networks:
    get:
      tags:
      - "networks"
      summary: "Lists all networks."
      operationId: "getNetworks"
      schemes:
      - "https"
      parameters: []
      responses:
        "200":
          description: ""
          schema:
            type: "array"
            items:
              $ref: "#/definitions/Network"
    post:
      tags:
      - "networks"
      summary: "Create a network."
      operationId: "createNetwork"
      schemes:
      - "https"
      parameters:
      - in: "body"
        name: "network"
        description: "Network"
        required: false
        schema:
          $ref: "#/definitions/Network"
        x-go-name: "Network"
        x-exportParamName: "Network"
      responses:
        "200":
          description: ""
          schema:
            $ref: "#/definitions/Network"
  /api/networks/{networkname}:
    get:
      tags:
      - "networks"
      summary: "Get a network."
      operationId: "getNetwork"
      schemes:
      - "https"
      parameters:
      - name: "networkname"
        in: "path"
        description: "name: network name"
        required: true
        type: "string"
        x-go-name: "Networkname"
        x-exportParamName: "Networkname"
      responses:
        "200":
          description: ""
          schema:
            $ref: "#/definitions/Network"
    put:
      tags:
      - "networks"
      summary: "Update pro settings for a network."
      operationId: "updateNetwork"
      schemes:
      - "https"
      parameters:
      - name: "networkname"
        in: "path"
        description: "name: network name"
        required: true
        type: "string"
        x-go-name: "Networkname"
        x-exportParamName: "Networkname"
      - in: "body"
        name: "network"
        description: "Network"
        required: false
        schema:
          $ref: "#/definitions/Network"
        x-go-name: "Network"
        x-exportParamName: "Network"
      responses:
        "200":
          description: ""
          schema:
            $ref: "#/definitions/Network"
    delete:
      tags:
      - "networks"
      summary: "Delete a network.  Will not delete if there are any nodes that belong\
        \ to the network."
      operationId: "deleteNetwork"
      schemes:
      - "https"
      parameters:
      - name: "networkname"
        in: "path"
        description: "name: network name"
        required: true
        type: "string"
        x-go-name: "Networkname"
        x-exportParamName: "Networkname"
      responses:
        "200":
          description: ""
          schema:
            $ref: "#/definitions/SuccessResponse"
...
definitions:
...
  Network:
    type: "object"
    properties:
      addressrange:
        type: "string"
        x-go-name: "AddressRange"
      addressrange6:
        type: "string"
        x-go-name: "AddressRange6"
      allowmanualsignup:
        type: "string"
        x-go-name: "AllowManualSignUp"
      defaultacl:
        type: "string"
        x-go-name: "DefaultACL"
      defaultinterface:
        type: "string"
        x-go-name: "DefaultInterface"
      defaultkeepalive:
        type: "integer"
        format: "int32"
        x-go-name: "DefaultKeepalive"
      defaultlistenport:
        type: "integer"
        format: "int32"
        x-go-name: "DefaultListenPort"
      defaultmtu:
        type: "integer"
        format: "int32"
        x-go-name: "DefaultMTU"
      defaultpostdown:
        type: "string"
        x-go-name: "DefaultPostDown"
      defaultudpholepunch:
        type: "string"
        x-go-name: "DefaultUDPHolePunch"
      isipv4:
        type: "string"
        x-go-name: "IsIPv4"
      isipv6:
        type: "string"
        x-go-name: "IsIPv6"
      netid:
        type: "string"
        x-go-name: "NetID"
      networklastmodified:
        type: "integer"
        format: "int64"
        x-go-name: "NetworkLastModified"
      nodelimit:
        type: "integer"
        format: "int32"
        x-go-name: "NodeLimit"
      nodeslastmodified:
        type: "integer"
        format: "int64"
        x-go-name: "NodesLastModified"
    description: "Network Struct - contains info for a given unique network\nAt  some\
      \ point, need to replace all instances of Name with something else like  Identifier"
    example:
      defaultinterface: "defaultinterface"
      isipv6: "isipv6"
      netid: "netid"
      defaultpostdown: "defaultpostdown"
      nodeslastmodified: 2
      addressrange: "addressrange"
      nodelimit: 5
      networklastmodified: 5
      defaultudpholepunch: "defaultudpholepunch"
      allowmanualsignup: "allowmanualsignup"
      defaultkeepalive: 0
      addressrange6: "addressrange6"
      defaultacl: "defaultacl"
      defaultlistenport: 6
      isipv4: "isipv4"
      defaultmtu: 1
    x-go-package: "github.com/gravitl/netmaker/models"
...
  SuccessResponse:
    type: "object"
    properties:
      Code:
        type: "integer"
        format: "int64"
      Message:
        type: "string"
      Response: {}
    title: "SuccessResponse is struct for sending error message with code."
    example:
      Response: ""
      Message: "Message"
      Code: 0
    x-go-package: "github.com/gravitl/netmaker/models"
...