# swagger/model_network.go
/*
 * Netmaker
 *
 * # API Usage  Most actions that can be performed via API can be performed via UI. We recommend managing your networks using the official netmaker-ui project. However, Netmaker can also be run without the UI, and all functions can be achieved via API calls. If your use case requires using Netmaker without the UI or you need to do some troubleshooting/advanced configuration, using the API directly may help.  # Authentication  API calls must be authenticated via a header of the format -H “Authorization: Bearer <YOUR_SECRET_KEY>” There are two methods to obtain YOUR_SECRET_KEY: 1. Using the masterkey. By default, this value is “secret key,” but you should change this on your instance and keep it secure. This value can be set via env var at startup or in a config file (config/environments/< env >.yaml). See the [Netmaker](https://docs.netmaker.org/index.html) documentation for more details. 2. Using a JWT received for a node. This can be retrieved by calling the /api/nodes/<network>/authenticate endpoint, as documented below.
 *
 * API version: 0.24.0
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

// Network Struct - contains info for a given unique network At  some point, need to replace all instances of Name with something else like  Identifier
type Network struct {
	Addressrange string `json:"addressrange,omitempty"`
	Addressrange6 string `json:"addressrange6,omitempty"`
	Allowmanualsignup string `json:"allowmanualsignup,omitempty"`
	Defaultacl string `json:"defaultacl,omitempty"`
	Defaultinterface string `json:"defaultinterface,omitempty"`
	Defaultkeepalive int32 `json:"defaultkeepalive,omitempty"`
	Defaultlistenport int32 `json:"defaultlistenport,omitempty"`
	Defaultmtu int32 `json:"defaultmtu,omitempty"`
	Defaultpostdown string `json:"defaultpostdown,omitempty"`
	Defaultudpholepunch string `json:"defaultudpholepunch,omitempty"`
	Isipv4 string `json:"isipv4,omitempty"`
	Isipv6 string `json:"isipv6,omitempty"`
	Netid string `json:"netid,omitempty"`
	Networklastmodified int64 `json:"networklastmodified,omitempty"`
	Nodelimit int32 `json:"nodelimit,omitempty"`
	Nodeslastmodified int64 `json:"nodeslastmodified,omitempty"`
}